modify api details discussed with WJ without JWT
