+++ backend/app/__init__.py
@@ -2,4 +2,4 @@ from flask import Flask
 from flask_cors import CORS
 app = Flask(__name__)
 CORS(app)
-from . import views
+from . import project
